+++ Sources/TuistGenerator/Linter/GraphLinter.swift
@@ -44,7 +44,6 @@ public class GraphLinter: GraphLinting {
         issues.append(contentsOf: lintDependencies(graphTraverser: graphTraverser))
         issues.append(contentsOf: lintMismatchingConfigurations(graphTraverser: graphTraverser))
         issues.append(contentsOf: lintWatchBundleIndentifiers(graphTraverser: graphTraverser))
-        issues.append(contentsOf: lintBundleIdentifiers(graphTraverser: graphTraverser))
 
         return issues
     }
@@ -304,47 +303,6 @@ public class GraphLinter: GraphLinting {
         }
 
         return foundIssues
-    }
-
-    private struct BundleIdKey: Hashable {
-        let bundleId: String
-        let platform: Platform
-    }
-
-    private func lintBundleIdentifiers(graphTraverser: GraphTraversing) -> [LintingIssue] {
-        var bundleIds = [BundleIdKey: [String]]()
-        graphTraverser.targets
-            .flatMap { $0.value.map(\.value) }
-            .forEach { target in
-                // skip duplicate check for bundle Ids that contain variables
-                // e.g.
-                // - `${MY_BUNDLE_ID}`
-                // - `prefix.${PRODUCT_NAME:rfc1034identifier}`
-                // - `${PRODUCT_NAME:rfc1034identifier}.suffix`
-                if target.bundleId.contains("$") {
-                    return
-                }
-
-                let key = BundleIdKey(bundleId: target.bundleId, platform: target.platform)
-
-                var targetsWithThisBundleId = bundleIds[key] ?? []
-                targetsWithThisBundleId.append(target.name)
-                bundleIds[key] = targetsWithThisBundleId
-            }
-
-        return duplicateBundleIdLintingIssue(for: bundleIds)
-    }
-
-    private func duplicateBundleIdLintingIssue(for targets: [BundleIdKey: [String]]) -> [LintingIssue] {
-        targets.compactMap { bundleIdKey, targetNames in
-            guard targetNames.count > 1 else {
-                return nil
-            }
-
-            let reason =
-                "The bundle identifier '\(bundleIdKey.bundleId)' is being used by multiple targets: \(targetNames.sorted().listed())."
-            return LintingIssue(reason: reason, severity: .warning)
-        }.sorted(by: { $0.reason < $1.reason })
     }
 
     struct LintableTarget: Equatable, Hashable {